Revert "Move context provider to avoid constant user revalidations"
This reverts commit e9515ad21b23e343895e1927ee6a3dce350c4654.
