+++ src/components/Widgets/FeedbackWidget/views/SubmittedView.js
@@ -20,8 +20,9 @@ const StyledHeading = styled(Subtitle)`
 `;
 
 const SubmittedView = () => {
-  const { abandon, selectedRating } = useFeedbackContext();
+  const { abandon } = useFeedbackContext();
   const { isMobile } = useScreenSize();
+  const { selectedRating } = useFeedbackContext();
   const shouldShowSupportLink = selectedRating <= 3;
 
   return (
