+++ src/layouts/preview-layout.js
@@ -155,7 +155,6 @@ const DefaultLayout = ({
           headingNodes={page?.options?.headings}
           selectors={page?.options?.selectors}
           isAssociatedProduct={isAssociatedProduct}
-          pageTitle={pageTitle}
         >
           <GlobalGrid>
             <PreviewHeader sidenav={sidenav} />
